tests: move test helpers out of docs_test.go
Move test helpers (which are also used in fish_test.go) out of
docs_test.go and into the existing helpers_test.go.

This will allow to use build tags to exclude features exposed by
docs.go.
